backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pymysql==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
email-validator==2.2.0
reportlab==4.4.3
qrcode[pil]==8.2
Pillow==11.3.0
orjson==3.11.3
httpx==0.28.1
alembic==1.16.5
slowapi==0.1.9
pytest==8.4.1
